SPI 정보를 입력할 DTO 클래스 정의, 주석 작성

+++ app/src/main/java/kr/djspi/pipe01/dto/SpiDepth.java
@@ -6,5 +6,8 @@ import lombok.Data;
 public class SpiDepth extends SpiData {
 
     private int id;
-    private int depth;
+    /**
+     * 관로 심도(단위 m)
+     */
+    private double depth;
 }
